move FieldFormat class outside of kibana plugin
