+++ src/core_plugins/kibana/common/field_formats/types/ip.js
@@ -1,4 +1,4 @@
-import { FieldFormat } from '../field_format';
+import { FieldFormat } from '../../../../../ui/field_formats/field_format';
 
 export class IpFormat extends FieldFormat {
   _convert(val) {
