+++ src/core_plugins/kibana/common/field_formats/types/string.js
@@ -1,5 +1,5 @@
 import { asPrettyString } from '../../utils/as_pretty_string';
-import { FieldFormat } from '../field_format';
+import { FieldFormat } from '../../../../../ui/field_formats/field_format';
 import { shortenDottedString } from '../../utils/shorten_dotted_string';
 
 export class StringFormat extends FieldFormat {
