+++ src/ui/field_formats/__tests__/field_format.js
@@ -1,6 +1,6 @@
 import _ from 'lodash';
 import expect from 'expect.js';
-import { asPrettyString } from '../../utils/as_pretty_string';
+import { asPrettyString } from '../../../core_plugins/kibana/common/utils/as_pretty_string';
 import { FieldFormat } from '../field_format';
 
 describe('FieldFormat class', function () {
