+++ src/ui/field_formats/content_types.js
@@ -1,6 +1,6 @@
 import _ from 'lodash';
-import { asPrettyString } from '../utils/as_pretty_string';
-import { getHighlightHtml } from '../highlight/highlight_html';
+import { asPrettyString } from '../../core_plugins/kibana/common/utils/as_pretty_string';
+import { getHighlightHtml } from '../../core_plugins/kibana/common/highlight/highlight_html';
 
 const types = {
   html: function (format, convert) {
